requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
coverage==7.8.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pprintpp==0.4.0
pycountry==24.6.1
pycountry-convert==0.7.2
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
repoze.lru==0.7
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
wheel==0.46.1
